Add --report; simplify CLI to one file
