--- scooby/cli/__init__.py
@@ -1,2 +1,0 @@
-"""Provide command-line interface (CLI)."""
-from scooby.cli.main import main  # noqa
